+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.9;
-const unsigned char AutoSDKVersionString[] = "1.9.0";
+double AutoSDKVersionNumber = 1.10;
+const unsigned char AutoSDKVersionString[] = "1.10.0";